moved the ProfilePic style component to different folder, changed updated import path and imported Api home component
